refactor(web): share the resend-code button across the one-time-code surfaces
The resend control (disabled while the cooldown runs, remaining seconds
in the label) was implemented three times — once as a local component in
the email-code sign-in page, twice inlined in the forgot-password and
verify-email pages. One shared component replaces all three; no behavior
change, the accessible name the tests read is unchanged.
